Improve test fixtures management. (#140)
