+++ tests/annuaire/test_annuaire.py
@@ -3,8 +3,6 @@ import pytest
 from django.test import Client
 from django.urls import reverse
 
-from tests.fixtures import data
-
 
 @pytest.fixture
 def client():
